This version is not working yet! I'm commiting it for reference because the ractor structure will remain mostly the same.
Major Changes:
Added the rest of the major functionality to the ModeMachine reactor.
Added guards to the USM that keeps the motor pins disabled based on a flag from the ModeMachine.
Added guards to the MotorController reactor to enable/disable motor commands based on a flag from ModeMachine.
Added guards to the HomeController reactor to enable/disable homing velocity commands based on a flag from ModeMachine.
Added extra possible modes to the MODE_ENUM to distinguish between RUN with position control or RUN with velocity control.

TODO:
Update ModeMachine to not try to trigger output changes inside of methods. Have internal copies updated inside of methods and have a reactor that sets the actual outputs to match the internal copies.
Test if the mode changes are triggered correctly from the ROS side.